backend/requirements-dev.txt
Django==3.1.4
djangorestframework==3.12.2
django-filter==2.4.0
django-map-widgets==0.3.0
psycopg2-binary==2.8.6
django-storages==1.11
docutils==0.16
jmespath==0.10.0
Pillow==8.0.1
python-dateutil==2.8.1
sqlparse==0.4.1
pytz==2020.4
s3transfer==0.3.3
urllib3==1.26.2
six==1.15.0
botocore==1.19.42
boto3==1.16.42